Add files via upload
The purpose of this report is to analyze the Telco dataset to understand the characteristics of the company's customers and to predict their income and churn status. This analysis aims to help the telecommunications company reduce its dependency on external vendors for income data by developing an internal predictive model. The report includes sample characteristics, multiple linear regression analysis, hypothesis testing, and interpretation of the results.